+++ src/notifiers/textbelt.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use textbelt::TextbeltClient;
-use log::{info,debug,error};
+use log::{info,debug};
 
 #[derive(Clone)]
 pub struct TextbeltSink {
@@ -8,10 +8,9 @@ pub struct TextbeltSink {
     api_key: String,
 }
 
-
-
 impl TextbeltSink {
     pub fn new(api_key: String, phone: String) -> Self {
+        debug!("Creating textbelt client!");
         Self { phone, api_key }
     }
     pub async fn send(&self, msg: &str) -> Result<()> {