+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
 ## v0.0.3
-`2025/09/09`
-Added new notifier (SMS)
+
+### Date
+`2025/09/08`
+
+### Summary
+Added new notifier (SMS) from [Textbelt](https://textbelt.com/) API.
+
 ## v0.0.2
 
 ### Date