+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.0.3
+
+### Date
+`2025/09/08`
+
+### Summary
+Added new notifier (SMS) from [Textbelt](https://textbelt.com/) API.
+
 ## v0.0.2
 
 ### Date